renewal category plugin
* show only one category in a category page
* delete perild feature
* delete category_icon feature
* optimize rebuilding index
